requirement.txt
streamlit
pandas
requests
python-dotenv
tqdm
scikit-learn
